move params into param file, add tests
