+++ x/ibchooks/keeper/params.go
@@ -1,0 +1,24 @@
+package keeper
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/provenance-io/provenance/x/ibchooks/types"
+)
+
+// GetParams returns the total set of the module's parameters.
+func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
+	store := ctx.KVStore(k.storeKey)
+	bz := store.Get(types.IbcHooksParamStoreKey)
+	if bz == nil {
+		return types.DefaultParams()
+	}
+	k.cdc.MustUnmarshal(bz, &params)
+	return params
+}
+
+// SetParams sets the module's parameters with the provided parameters.
+func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
+	store := ctx.KVStore(k.storeKey)
+	bz := k.cdc.MustMarshal(&params)
+	store.Set(types.IbcHooksParamStoreKey, bz)
+}
